Refactor game logic and fix prompt handling issues
- Consolidate all game logic into useGameLogic hook
- Fix duplicate miss handling and auto-miss timers
- Add proper feint support with validation
- Improve prompt completion logic to handle feints correctly
- Add haptic feedback and audio utilities
- Refactor GameScreen for better readability
- Add proper UI components (GameHUD, GameInputArea, PreRoundDisplay)
- Fix swipe gesture handling and visual feedback
- Add proper opponent data integration
- Remove console logs and improve performance
