+++ src/utils/hapticUtils.ts
@@ -1,0 +1,33 @@
+import * as Haptics from 'expo-haptics';
+import { HapticType } from '../types/game';
+
+// ============================================================================
+// HAPTIC UTILITIES
+// ============================================================================
+
+export const triggerHaptic = (type: HapticType) => {
+  try {
+    switch (type) {
+      case 'light':
+        Haptics.impactAsync(Haptics.ImpactFeedbackStyle.Light);
+        break;
+      case 'medium':
+        Haptics.impactAsync(Haptics.ImpactFeedbackStyle.Medium);
+        break;
+      case 'heavy':
+        Haptics.impactAsync(Haptics.ImpactFeedbackStyle.Heavy);
+        break;
+      case 'success':
+        Haptics.notificationAsync(Haptics.NotificationFeedbackType.Success);
+        break;
+      case 'error':
+        Haptics.notificationAsync(Haptics.NotificationFeedbackType.Error);
+        break;
+      case 'warning':
+        Haptics.notificationAsync(Haptics.NotificationFeedbackType.Warning);
+        break;
+    }
+  } catch (error) {
+    console.log('Haptic error:', error);
+  }
+};
